#106 - Querydsl -> Jdsl 리팩토링
- Querydsl 의존성 제거
- limit dsl 추가
- 비효율적인 부분 제거

+++ src/main/kotlin/com/sat/board/domain/dto/query/ArticleWithCount.kt
@@ -6,7 +6,7 @@ data class ArticleWithCount(
     val id: Long,
     val title: String,
     val category: String,
-    val commentCount: Int,
-    val likeCount: Int,
+    val commentCount: Long,
+    val likeCount: Long,
     val createdDateTime: LocalDateTime,
 )
